Securisation de l'inclusion des variables gloables Ajout du test d'intégrité de l'arborescence des dossiers

+++ scripts/reboot_rtorrent.sh
@@ -1,7 +1,14 @@
 #!/bin/bash
 
+
+#verification qu'il y a bien une variable passée en paramètre
+if [ -z $1 ]
+then
+        echo "Aucun utilisateur spécifié, arrêt de reboot_rtorrent.sh!"
+        exit
+fi
+
 killall --user $1 rtorrent
 rm /home/$1/.session/rtorrent.lock
 su --command="screen -S $1-rtorrent -X quit" $1
-
 su --command="screen -dmS $1-rtorrent rtorrent" $1
